requirements.txt
asgiref==3.5.2
Django==4.0.6
django-autoslug==1.9.8
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytils==0.4.1
sqlparse==0.4.2
tzdata==2022.1
